Added 'dedup' command to remove all duplicates in one folder from another.
This works by reading in a catalog file from both folders, finding all
the files in the "to check" folder that are identical to one or more
files in the "base" folder.

This introduces FileHash which records a single file on disk and its
hash+filesize+relativepath and a FileCatalog which is a way to group all
file hashes in a folder by hash (so all files that are identical are in
the same list together).

This also updates the delete of both 'dedup' and 'search' to fail more
gracefully: keeping track of how many it successfully deleted and
printing it even on failure.

Todo in the future, be much more careful about deleting files. Make sure
the hashes of both source and destination are still accurate and match
each other.