Deleted KFileBackup because it has been obsoleted by KCatalog.
KCatalog does largely the same thing while being "simpler."

KFileBackup relied on a central database for keeping track of backed up
file, their hashes, their source locations, and the backup media. This
put a lot of emphasis on this central DB with no clear way to version it
or backup this DB.

KCatalog on the other hand doesn't even try to solve the problem of
tracking backed-up files in a centralized way. Instead it works by
simply "cataloging" a given directory and storing that cached data in
that very same directory. It then has tools for reading these caches and
operating on them (e.g. comparing files, de-duplicating, organizing,
etc.). When a physical backup media is created, it can be cataloged and
the catalog cache can then be burned onto that media and also kept
locally (caches are snapshots in time and a snapshot of read-only media
means an immutable cache, therefore there is no need to try to backup or
version the catalog files).