KCatalog: Reorganized and reworded all commands to be clearer.
The commands were previously not as clear and I always forgot what they
did and how they worked and which ones I was supposed to use (even
though I wrote them). This commit groups the commands under "catalog"
and "dir" commands to be clear which ones care about catalog files and
which ones are just helpers. I changed some of the catalog commands to
operate on a catalog file to be clear that the file system files are
ignored and you must make sure your catalog is up-to-date by using
catalog-create.

This is completely untested. Up next I hope to write unit tests for all
this super critical archival and cataloging logic.