KCatalog: Cleaned up data types and type names for clarity.
This makes the types names easier to understand so reading the commands
is actually useful.

A 'Catalog' has a list of 'FileInstance' objects which contain a
relative path, a file size, and a 'Hash256' for the file contents hash.

This is completely untested. Still hoping to get to unit tests.