Added automatic version numbering to build output (with dirty changes detection).
This commit adds rich version numbers to both KCatalog and
KPhotoOrganizer (and serves as a framework for any VS project stored in
Git). When a project is built the output is automatically versioned with
the Git commit hash and the base version number defined in version.txt.
"Dirty changes detection" in this case means that we check to see if the
build is including any non-committed changes and, if so, appends
'-dirty' to the file version. This way every build can easily be traced
back to a specific git commit to know what source code was running and
if the build is "unofficial" by having "dirty" or non-committed changes
(in which case you know that you can't trust that build).

This works by shelling out to Git twice: once for 'git status' to detect
if there are any changes (i.e. a dirty build off of non-committed
changes) and then for getting the hash of the commit we are on. These
two pieces of information go into the 4th digit of the AssemblyVersion
and AssemblyFileVersion. Then there is a version.txt which defines the
major minor and build numbers (the first two digits). The third digit is
always zero right now.

1.2.0.a4cd53-dirty
| |   |    | |   |
 ^ the fixed "base" version defined in version.txt
         ^ the commit hash of the source code that was built
	       ^ whether or not there are non-committed changes in this build
	         (if this is a straight build of a commit this is ommitted)

This version number will also be placed into every catalog file created
by KCatalog to make sure we know exactly what source code created that catalog file.