KCatalog: Added photo archival commands photo-archive and photo-archive-validate.
photo-archive will automatically take a bunch of photos and videos in a
folder and move them to be archived with a very specific folder structure:
Archive/
  XXXX (year)/
    XX-yyyyyy (month number and name)/
      MM-DD-YYYY (month, day, year)/
        files.ext

photo-archive-validate just makes sure there is nothing wrong with the
archive and its files.

These commands replace the separate KPhotoOrganizer application so that
has been deleted.