KCatalog: Added the 'catalog-compare-unique' command to find unique files that are _not_ in both catalogs.
This command is different from the old 'catalog-compare' in that it
looks for non-duplicated files, whereas 'catalog-compare' only reported
exact duplicates.

As part of this I renamed 'catalog-compare' to 'catalog-compare-duplicates'
to make it more clear vs 'catalog-compare-unique'. I always refactored
the code and tests to match this naming. This applies to the directory
commands as well: 'dir-compare' was renamed to 'dir-compare-duplicates'.