KCatalog: Made photo-archive work only with an existing catalog.
Now, archiving photos requires an existing catalog in place, and a
photo will not be archived if its hash is already in the catalog
somewhere. This makes the workflow of finding old photos significantly
easier: just update the catalog, archive the photos, and the ones left
behind are probably already archived elsewhere (but check the logs).

This means catalog-create has to be done on the destination directory
first. The tests were updated to do this.

Also removed some incorrect documentation about automatic VLC
conversions which do not happen at all.