KCatalog: catalog-update now logs if a file was "moved".
Instead of the log saying "removed" and then "added", it now detects if
a file was moved and says as much. This is much better for
understanding what has happened but it separates the logging from the
logic a little bit (all the "removed" logs are spit out after all the
removals were actually done).