KCatalog: catalog-update now detects duplicate file removal and bug fix.
catalog-update had a bug introduced by the "moved" detection commit
where if multiple files were removed the Dictionary.Add would throw an
exception.  This is now guarded with a Dictionary.ContainsKey check so
we only add the first file we come across of that hash, which should be
the oldest because of file path sorting (for photo archives anyway).

Also added duplicate file removal detection. This works well with the
new catalog-compare-duplicates-self command so you know that the files
you removed from the catalog are duplicates when you do catalog-update.