KCatalog: photo-archive now handles UTC timestamped file names of photos.
KCatalog photo archiving was designed around the idea that a file name
has a timestamp and that this timestamp was in local time. Google Pixel
phones instead use UTC time in the file name timestamp which breaks
KCatalog photo archiving since photos can go into the wrong date folders
(if the time is close enough to midnight to roll over to the next day,
within 4-5 hours depending on DST).

In addition to making sure the photos get organized into the correct
date folders, photo archiving via 'photo-archive' now automatically
fixes these UTC timestamp file names so the file names are in local time
instead.

This bumped KCatalog to version 1.2 since it is now handling a new
ecosystem (Pixel phones) and requires a more rigid timestamp file name
format (see below) that is technically a compatibility break.

I verified that the DateTime struct correctly handles converting a UTC
time to local time for both DST and non-DST date ranges. In other words,
DateTime does _not_ use "today's" DST status when doing ToLocalTime(),
but will instead detect if _that date_ would be under DST and adjusts
accordingly.

WARNING: This does _not_ detect the local time of the photo, it assumes
that the local time of the executing PC is the same as the photo (which
it usually will be). This is an ok assumption for me personally but may
bite me when traveling to different time zones. A more robust solution
would be to either get the timestamp out of the metadata (which should
be in local time) or get the local time zone information out of the
metadata instead of assuming the time zone of the executing PC. However,
this won't work for mp4 files (as far as I can tell they store UTC
timestamps with no timezone) and it would introduce a complex and slow
dependency into photo archiving that I would rather avoid. So I accept
this risk.

As part of this, the regular expression regex for matching timestamped
file names was made more forceful: the millisecond is no longer optional
and the suffix capture was made non-lazy to actually get the remaining
text. These changes were important so we can change the timestamp of a
UTC file while preserving everything else.

This commit adds the PhotoArchive_UtcPhotos test to make sure that UTC
timestamped file names work both during DST and outside of DST. It also
tests that suffixes and the rest of the filename is preserved.

This also makes a few other changes:
-Fixed the flaky CatalogUpdateA_Changes test by introducing a sleep to
guarantee that created date and modified date of the catalog are
different (sometimes the test would run so fast the milliseconds were
the same for created and updated).
-Removed an example file name from the PhotoArchive test that did not
have a millisecond value in its file name time stamp (since we force
that milliseconds exist now).