KCatalog: catalog-create and catalog-update ignore .git folders.
The .git folder is a special folder where all of the internals of a Git
repository reside. It may be desirable to use both KCatalog and a Git
repository for the same set of files, but the KCatalog files shouldn't
try to keep track of the state of the internal Git repo. So this commit
makes it such that the .git folder is special and will be ignored by
KCatalog.