KCatalog: Fixed bug where catalog-update wouldn't work because of a trailing slash.
catalog-update would write a .kcatalog file that had a trailing slash at
the end of the directory path for BaseDirectoryPath. I worked around
this in the past by manually editing the .kcatalog file after creating
it to remove the trailing slash. If I didn't then catalog-update would
fail saying that the .kcatalog file you want to update isn't in the same
directory it was created in (because the trailing slash throws off the
string comparison). Now that catalog-create does not add a trailing
slash, catalog-update works just fine.

The trailing slash was added as part of some deliberate code during
command line argument parsing that enforced that a directory path always
ends with a trailing slash. I don't know why. Removing this trailing
slash enforcement might break some of the other commands, but all of the
unit tests currently pass.

Why didn't the unit tests catch this? It appears the mocking library has
a difference in behavior for mock directories vs real directories. The
code that adds the trailing slash to user input happens for both real
and mock directories, but the mock directories do _not_ preserve the
trailing slash. In other words, the mock directories will remove the
trailing slash but real directories won't. This is why the tests
succeeded but real kcatalogs failed.