KCatalog: Added the 'catalog-check-file-integrity-from-backup' command to see if files on disk have become corrupt compared to what was backed up.
This works by looking at all files in the current catalog and finding
that same file in a backed up catalog by relative path. If a matching
file is found the hashes are checked, if the hashes are different the
file must be modified and it is listed. This only works if you update
your catalog first to detect any changes on disk. This will also print
out any renames though that's usually less interesting and more just
noise for you to parse through.